Initial restructuring of files.
