+++ src/advanced-search/state/reducers.js
