+++ src/footer/UniProtFooter.jsx
