--- src/ui/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import query from './query';
-// import namespace from './namespace';
-// import searchTerms from './searchTerms';
-
-export default combineReducers({
-  query,
-  // namespace,
-  // searchTerms,
-});
