--- src/ui/utils/index.js
@@ -1,3 +1,0 @@
-export { default as createQueryString } from './query';
-export { default as fetchData } from './fetchData';
-export { default as createEmptyClause } from './clause';
